feat(cache): cache options array
the `orphang_indicator_options` function is called multiple times for both Admin Panel and normal page frontend. This causes the plugins to generate an array of options many times. As this array consists of post types, taxonomies and installed plugins it takes around 140ms of a load time of the webpage. With this cache this value is reduced to 1ms